moved review file to app/models
